--- src/main/java/cart/configuration/resolver/Checkout.java
@@ -1,11 +1,0 @@
-package cart.configuration.resolver;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Checkout {
-}
